hw4/requirements.txt
flask
requests 
line-bot-sdk 
mysql-connector-python